fix issue #6
rules-io nodes on different flows determine input and output key value pairs from different nodes. This bug is a showstopper. The reason for this behavior is a wrong file name in the node. In this case comes topics to the inputs, there are unknown for the rules. This is the real bug of this behavior. By unknown topics are not outputs allowed.

The function getValidateInputTopic is implemented. This functio compare the input topic with all possivle topics from the rules file. When the income topic is part of the config section of the rules file the node runs to the next step. Otherwise the node stops and return null.